Move extractors to its own files
