+++ yt_dlp/extractor/_extractors.py
@@ -285,6 +285,7 @@ from .cammodels import CamModelsIE
 from .camsoda import CamsodaIE
 from .camtasia import CamtasiaEmbedIE
 from .camwithher import CamWithHerIE
+from .canal1 import Canal1IE
 from .canalalpha import CanalAlphaIE
 from .canalplus import CanalplusIE
 from .canalc2 import Canalc2IE
@@ -294,6 +295,7 @@ from .canvas import (
     VrtNUIE,
     DagelijkseKostIE,
 )
+from .caracoltv import CaracolTvPlayIE
 from .carambatv import (
     CarambaTVIE,
     CarambaTVPageIE,
@@ -1039,8 +1041,6 @@ from .mediasite import (
 )
 from .mediastream import (
     MediaStreamIE,
-    Canal1IE,
-    CaracolTvPlayIE,
     WinSportsVideoIE,
 )
 from .mediaworksnz import MediaWorksNZVODIE
