+++ yt_dlp/extractor/mediastream.py
@@ -1,14 +1,9 @@
-import base64
-import itertools
-import json
 import re
-import uuid
 
 from .common import InfoExtractor
 from ..utils import (
     clean_html,
     remove_end,
-    js_to_json,
     traverse_obj,
     urljoin,
 )
@@ -144,161 +139,6 @@ class MediaStreamIE(MediaStreamBaseIE):
             'is_live': player_config.get('type') == 'live',
             'thumbnail': self._og_search_thumbnail(webpage),
         }
-
-
-class CaracolTvPlayIE(InfoExtractor):
-    _VALID_URL = r'https?://play\.caracoltv\.com/videoDetails/(?P<id>[^/?]+)'
-    _NETRC_MACHINE = 'caracoltv-play'
-
-    _TESTS = [{
-        'url': 'https://play.caracoltv.com/videoDetails/OTo4NGFmNjUwOWQ2ZmM0NTg2YWRiOWU0MGNhOWViOWJkYQ==',
-        'info_dict': {
-            'id': 'OTo4NGFmNjUwOWQ2ZmM0NTg2YWRiOWU0MGNhOWViOWJkYQ==',
-            'title': 'La teoría del promedio',
-            'description': 'md5:1cdd6d2c13f19ef0d9649ab81a023ac3',
-        },
-        'playlist_count': 6,
-    }, {
-        'url': 'https://play.caracoltv.com/videoDetails/OTo3OWM4ZTliYzQxMmM0MTMxYTk4Mjk2YjdjNGQ4NGRkOQ==/ella?season=0',
-        'info_dict': {
-            'id': 'OTo3OWM4ZTliYzQxMmM0MTMxYTk4Mjk2YjdjNGQ4NGRkOQ==',
-            'title': 'Ella',
-            'description': 'md5:a639b1feb5ddcc0cff92a489b4e544b8',
-        },
-        'playlist_count': 10,
-    }, {
-        'url': 'https://play.caracoltv.com/videoDetails/OTpiYTY1YTVmOTI5MzI0ZWJhOGZiY2Y3MmRlOWZlYmJkOA==/la-vuelta-al-mundo-en-80-risas-2022?season=0',
-        'info_dict': {
-            'id': 'OTpiYTY1YTVmOTI5MzI0ZWJhOGZiY2Y3MmRlOWZlYmJkOA==',
-            'title': 'La vuelta al mundo en 80 risas 2022',
-            'description': 'md5:e97aac36106e5c37ebf947b3350106a4',
-        },
-        'playlist_count': 17,
-    }]
-
-    _USER_TOKEN = None
-
-    def _extract_app_token(self, webpage):
-        config_js_path = self._search_regex(
-            r'<script[^>]+src\s*=\s*"([^"]+coreConfig.js[^"]+)', webpage, 'config_js_url', fatal=False)
-
-        mediation_config = {} if not config_js_path else self._search_json(
-            r'mediation\s*:', self._download_webpage(
-                urljoin('https://play.caracoltv.com/', config_js_path), None, fatal=False),
-            'mediation_config', None, transform_source=js_to_json, fatal=False)
-
-        key = traverse_obj(
-            mediation_config, ('live', 'key')) or '795cd9c089a1fc48094524a5eba85a3fca1331817c802f601735907c8bbb4f50'
-        secret = traverse_obj(
-            mediation_config, ('live', 'secret')) or '64dec00a6989ba83d087621465b5e5d38bdac22033b0613b659c442c78976fa0'
-
-        return base64.b64encode(f'{key}:{secret}'.encode()).decode()
-
-    def _login(self, email, password):
-        webpage = self._download_webpage('https://play.caracoltv.com/', None, fatal=False)
-        app_token = self._extract_app_token(webpage)
-
-        bearer_token = self._download_json(
-            'https://eu-gateway.inmobly.com/applications/oauth', None, data={}, headers={
-                'Authorization': f'Basic {app_token}',
-            })['token']
-
-        self._USER_TOKEN = self._download_json(
-            'https://eu-gateway.inmobly.com/user/login', None, headers={
-                'Content-Type': 'application/json',
-                'Authorization': 'Bearer' + bearer_token
-            }, data=json.dumps({
-                'device_data': {
-                    'device_id': str(uuid.uuid4()),
-                    'device_token': '',
-                    'device_type': 'web'
-                },
-                'login_data': {
-                    'enabled': True,
-                    'email': email,
-                    'password': password,
-                }
-            }).encode())['user_token']
-
-    def _perform_login(self, username, password):
-        self._login(username, password)
-
-    def _extract_series_season(self, series_id, season_id, season_number):
-        api_response = self._download_json(
-            'https://eu-gateway.inmobly.com/feed', series_id, query={'season_id': season_id},
-            headers={'Authorization': 'Bearer ' + self._USER_TOKEN})
-
-        for episode in api_response['items']:
-            formats, subtitles = self._extract_m3u8_formats_and_subtitles(episode['stream_url'], series_id, 'mp4')
-
-            yield {
-                'id': episode['id'],
-                'title': episode.get('name'),
-                'description': episode.get('description'),
-                'formats': formats,
-                'subtitles': subtitles,
-                'thumbnails': traverse_obj(
-                    episode, ('extra_thumbs', ..., {'url': 'thumb_url', 'height': 'height', 'width': 'width'})),
-                'series_id': series_id,
-                'season_id': season_id,
-                'season_number': season_number,
-                'episode_number': episode.get('item_order'),
-            }
-
-    def _real_extract(self, url):
-        series_id = self._match_id(url)
-
-        if self._USER_TOKEN is None:
-            self._login('guest@inmobly.com', 'Test@gus1')
-
-        series_api_response = self._download_json(
-            'https://eu-gateway.inmobly.com/feed', series_id, query={'include_ids': series_id},
-            headers={'Authorization': 'Bearer ' + self._USER_TOKEN})['items'][0]
-
-        return self.playlist_result(
-            itertools.chain.from_iterable(
-                self._extract_series_season(series_id, season['id'], season.get('order'))
-                for season in series_api_response['seasons']),
-            series_id, playlist_title=series_api_response.get('name'),
-            playlist_description=series_api_response.get('description'))
-
-
-class Canal1IE(InfoExtractor):
-    _VALID_URL = r'https?://(?:www\.|noticias\.)?canal1\.com\.co/(?:[^?&])+/(?P<id>[\w-]+)'
-
-    _TESTS = [{
-        'url': 'https://canal1.com.co/noticias/napa-i-una-cadena-de-produccion-de-arroz-que-se-quedo-en-veremos-y-abandonada-en-el-departamento-del-choco/',
-        'info_dict': {
-            'id': '63b39f6b354977084b85ab54',
-            'display_id': 'napa-i-una-cadena-de-produccion-de-arroz-que-se-quedo-en-veremos-y-abandonada-en-el-departamento-del-choco',
-            'title': 'Ñapa I Una cadena de producción de arroz que se quedó en veremos y abandonada en el departamento del Chocó',
-            'description': 'md5:bc49c6d64d20610ea1e7daf079a0d013',
-            'thumbnail': r're:^https?://[^?#]+63b39f6b354977084b85ab54',
-            'ext': 'mp4',
-        },
-    }, {
-        'url': 'https://noticias.canal1.com.co/noticias/tres-i-el-triste-record-que-impuso-elon-musk-el-dueno-de-tesla-y-de-twitter/',
-        'info_dict': {
-            'id': '63b39e93f5fd223aa32250fb',
-            'display_id': 'tres-i-el-triste-record-que-impuso-elon-musk-el-dueno-de-tesla-y-de-twitter',
-            'title': 'Tres I El triste récord que impuso Elon Musk, el dueño de Tesla y de Twitter',
-            'description': 'md5:d9f691f131a21ce6767ca6c05d17d791',
-            'thumbnail': r're:^https?://[^?#]+63b39e93f5fd223aa32250fb',
-            'ext': 'mp4',
-        },
-    }, {
-        # Geo-restricted to Colombia
-        'url': 'https://canal1.com.co/programas/guerreros-canal-1/video-inedito-guerreros-despedida-kewin-zarate/',
-        'only_matching': True,
-    }]
-
-    def _real_extract(self, url):
-        display_id = self._match_id(url)
-        webpage = self._download_webpage(url, display_id)
-
-        url = self._search_regex(r'"embedUrl"\s*:\s*"([^"]+)', webpage, "embedUrl")
-
-        return self.url_result(url, display_id=display_id, url_transparent=True)
 
 
 class WinSportsVideoIE(MediaStreamBaseIE):
